requirements/dev.txt
# --- Core Django ---
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
drf-nested-routers==0.93.4
# psycopg2==2.9.9   # Uncomment if using PostgreSQL in production

# --- Authentication ---
django-allauth==0.57.0
dj-rest-auth==2.2.8
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
cryptography==41.0.7

# --- HTTP Client ---
requests==2.31.0

# --- Google Auth ---
google-auth==2.35.0
